Expects a format on the lines of:
```
{
  "<url>": {
   "user": "user", //optional
   "password": "password", //optional
   "realm": "Realm of maven central, nexus", //optional
   "optional": false //optional,
   "headers": {"Authentation": "foobar"}, //optional
   "httpsOnly": true, // optional
   "passOnRedirect": true //optional
  }
}
```

Failing to parse will log a messsage to the console.
We will also output an info message if we dont find this file.
The file is expected to be in a directory according to the [XDG](https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html) spec, or
whatver is native to the current platform.