Improve contributor experience
Hi, thanks for this project. It's early days but I'm super excited
by the potential and it's working great for us so far.

When attempting to contribute, I checkout the source but I could
not run `bleep compile`, there were some errors regarding undefined
types, like `bleep-model@jvm213 not found: type BleepVersion`

I looked at the CI and saw that you run `bleep generate-resources`
before compiling, so I tried that.

I'm assuming this won't be required in the future, when things like
https://github.com/oyvindberg/bleep/issues/19 are resolved.