Drop 2.12, at least for now.
Context:
The reason the 2.12 support is there is ironically future-proofing. By having bleep cross-compiled to 2.12, it's possible to port sbt plugins without upgrading the scala version they are written in, and then use them in bleep scripts written in 2.12. I'm willing to drop this, since porting random 2.12 code to 2.13 is normally easy. Dropping 2.13 would not be acceptable yet, it's only recently that it was possible to port the sonatype integration code to 3.

Also I thought it beneficial in case it was needed to embed bleep in sbt or vice versa, but I don't see a path in that direction for now.