Added support for excluding tests.
The way this works is mutually exclusive with defining a subset of tests
that only should run.

It hooks into the same bsp API with the logic inverted. Instead of
saying run only these test that come from the user through the cli we
say run all tets besides the ones the user specified in the cli