Fix bug where file watching watches log file. Log events are logged, and we waste a lot of resources.
Solution is to move `.bleep/builds/bsp/project-selection.yaml` to `.bleep/conf/bsp-project-selection.yaml`.

The file watching works by watching the parent folder of a file. in this case that included all of `.bloop` with all class files and the log file.
by moving the project selection file out, we avoid this