--test-only (#394)
* wip: --test-only

* Initial support for specifying test suites when running tests

It is now possible to specify which test suites to run using the -o or
--only syntax.

The tests can be specified using the class name of the test suite or by
using the fully qualified name.

Example:
bleep test app-tests -o TestSuiteOne --only TestSuiteTwo
fully qualified:
bleep test app-tests -o com.foo.TestSuiteThree

* Updated test-only functionality for no matching tests

If no tests-suites in the project match the specified test suites
provided by the user no tests will run at all. If any of the tests
match, but not all the matching tests will run.

* Added support for excluding tests.

The way this works is mutually exclusive with defining a subset of tests
that only should run.

It hooks into the same bsp API with the logic inverted. Instead of
saying run only these test that come from the user through the cli we
say run all tets besides the ones the user specified in the cli

* Removed unnecessary calls to bloop

We only query bloop for the test suites if the -o or -x flags have been
passed to the CLI

* Cleanup

- move as much as possible to static scope (inside object) to it's testable
- I struggled to follow the include/exclude logic, so I tried to make it clearer
- allow combining `-o` and -`x`, latter takes precedence.
- "garbage collect" `target` names (that is, project ids) so bloop gets less work to do after filtering. don't ask it to run test projects we know we're not going to run test suites in

---------

Co-authored-by: Øyvind Raddum Berg <elacin@gmail.com>