Adds support for --link and --debug when linking with bloop
Bloop supports the client dynamically setting the linker mode using the
when passing the --link flag by passing an optional --debug or --release
flag. If --relase or --debug is not passed bloop will link with whatever
mode is set in bloop config

bump bloop-config to 2.2.0. Now published for Scala3

remove unused import