Removes linker mode from bleep model.
Bleep now always defaults linking to debug. If the user wants to link in
release mode the user has to pass the --release flag. This will cause
bloop to link in release mode.

The way this works is by defaulting the linker mode in bloop-config to
debug. Then we override the linker mode through bsp and set it to
release when the user passes the --relase (-r) flag when linking