Adds a link command (#445)
* Adds a link command

This is useful for projects where linking is a part of another toolchain such as
a JS project where we might link a ScalaJS project that uses JSExport and then
build the resulting files with e.g. vite.

* Adds support for --link and --debug when linking with bloop

Bloop supports the client dynamically setting the linker mode using the
when passing the --link flag by passing an optional --debug or --release
flag. If --relase or --debug is not passed bloop will link with whatever
mode is set in bloop config

bump bloop-config to 2.2.0. Now published for Scala3

remove unused import

* Removes linker mode from bleep model.

Bleep now always defaults linking to debug. If the user wants to link in
release mode the user has to pass the --release flag. This will cause
bloop to link in release mode.

The way this works is by defaulting the linker mode in bloop-config to
debug. Then we override the linker mode through bsp and set it to
release when the user passes the --relase (-r) flag when linking

* Fixes incompatability in VersionCombo for SN > 0.5

SN >0.5 uses a different versioning shcheme for the version combo than
previous versions. Now we add the scala version to the native library
version to get the correct version for the scala version

* Add support for release-mode for SN linker

SN can link release in 3 different modes. fast/size/full. This can now
be set in the bleep.yaml and is forwarded to bloop. It defaults to
release-fast if not set. This is only relevant when linking with
--release

* Add support for nativeLTO option for SN

* Add support for native multithreading option

* Add support for optimize, embedResources and incremental compilation

Add support for the latest supported configuration options in
bleep-config/bloop

* Fix encoding/decoding of NativeBuildTarget/LTO/LinkerReleaseMode